build(event-display): move XR managers to a separate directory
