+++ packages/phoenix-event-display/src/managers/three-manager/xr/vr-manager.ts
