+++ packages/phoenix-event-display/src/managers/three-manager/xr/xr-manager.ts
